feat(statsbox): added numbers to the boxes
moved the calculation of assets into home component and passing it down to the assetlists component
